5-ProtocolosComunicacion/ejercicio3/raspberry/requirements.txt
Flask==2.1.2
RPi.GPIO==0.7.1
